Create starting point from input
This code snippet lets a user select a starting town from a list of towns. It  prints a list of towns with numbers (1, 2, 3, etc.) so the user can select one.Takes user input and converts it to an integer.

Subtracts 1 because list indices start at 0, but the menu shows options starting at 1.Checks if the entered number is within the valid range (i.e., between 0 and the last index of town_names). The except block catches these errors and exits the program with an error message. If no error occurs, the selected town name is assigned to start.